+++ src/cli.ts
@@ -1,4 +1,5 @@
 import { ParserPlugin } from "@babel/parser";
+import chalk from "chalk";
 import program from "commander";
 import { cosmiconfigSync } from "cosmiconfig";
 import path from "path";
@@ -6,81 +7,70 @@ import { version } from "../package.json";
 import { ReportType } from "./api";
 import { print, printError, styleTitle } from "./printer";
 
-function listOption<T>(x: T, acc: T[] = []): T[] {
-  acc.push(x);
-  return acc;
-}
-
 program.name("jsx-info");
 program.version(version, "-v, --version");
 program
-  .arguments("[components...]")
   .description("Displays a report of JSX component and prop usage")
   .option("--no-progress", "do not show progress messages")
   // NOTE: chalk automatically reads process.argv for these flags, and commander
   // does not correctly handle allowing both "--foo" and "--no-foo" flags
   //
   // https://github.com/tj/commander.js/issues/108
+  .option("--components <components...>", "which components to scan")
   .option("--color", "force enable color")
   .option("--no-color", "force disable color")
   .option("--no-config", "disable config file")
   .option("--no-gitignore", "disable reading .gitignore files")
+  .option("--babel-plugins <plugins...>", "adds a Babel plugin")
   .option(
-    "--add-babel-plugin <PLUGIN>",
-    "adds a babel plugin (repeatable)",
-    listOption
-  )
-  .option(
-    "--directory <DIRECTORY>",
+    "--directory <directory>",
     "directory to use as the base for finding files instead of cwd"
   )
   .option(
-    "--ignore <PATTERN>",
-    "adds a glob pattern used to ignore input files (repeatable)",
-    listOption
+    "--ignore <patterns...>",
+    "adds a glob pattern used to ignore input files"
   )
   .option(
-    "--files <PATTERN>",
-    "glob pattern used to find input files (repeatable)",
-    listOption
+    "--files <patterns...>",
+    "glob pattern used to find input files (repeatable)"
   )
   .option("--report <usage|props|lines>", "specify which report to show")
   .option(
-    "--prop <PROP>",
+    "--prop <prop>",
     "which prop to search for when running a lines report"
   );
 
 program.on("--help", () => {
-  print(`
+  print(chalk`
 Examples:
   Use jsx-info interactively:
-      npx jsx-info
+      {bold.cyan npx jsx-info}
 
   Display info only for <div> and <Tab.Container>
-      npx jsx-info div Tab.Container
+      {bold.cyan npx jsx-info --components} {underline.magenta div Tab.Container}
 
   See lines where className prop was used on <div> component
-      npx jsx-info --report lines --prop className div
+      {bold.cyan npx jsx-info --report} {underline.magenta lines} {bold.cyan --prop} {underline.magenta className} {bold.cyan --components} {underline.magenta div}
 
   See lines where \`id\` prop was used on any component
-      npx jsx-info --report lines --prop id "*"
+      {bold.cyan npx jsx-info --report} {underline.magenta lines} {bold.cyan --prop} {underline.magenta id} {bold.cyan --components} {underline.magenta "*"}
 
   See lines where kind prop was used with value "primary" on Button component
-      npx jsx-info --report lines --prop kind=primary Button
+      {bold.cyan npx jsx-info --report} {underline.magenta lines} {bold.cyan --prop} {underline.magenta kind=primary} {bold.cyan --components} {underline.magenta Button}
 
   Ignore folders based on glob paths
-      npx jsx-info --ignore '**/__test__' --ignore packages/legacy
+      {bold.cyan npx jsx-info --ignore} {underline.magenta "**/__test__"} {underline.magenta packages/legacy}
 
   Enable Babel plugins
-      npx jsx-info --add-babel-plugin decorators-legacy --add-babel-plugin pipelineOperator
+      {bold.cyan npx jsx-info --babel-plugins} {underline.magenta decorators-legacy} {underline.magenta pipelineOperator}
 
   Example .jsx-info.json config file
-      {
-        "babelPlugins": ["decorators-legacy", "pipelineOperator"],
-        "directory": "src",
-        "ignore": ["**/__test__", "legacy/**"],
-        "files": ["**/*.{js,jsx,tsx}"]
-      }
+      ${"{"}
+        {cyan "babelPlugins"}: [{magenta "decorators-legacy"}, {magenta "pipelineOperator"}],
+        {cyan "directory"}: {magenta "src"},
+        {cyan "ignore"}: [{magenta "**/__test__"}, {magenta "legacy/**"}],
+        {cyan "files"}: [{magenta "**/*.${"{"}js,jsx,tsx${"}"}"}]
+      ${"}"}
 `);
 });
 
@@ -125,11 +115,11 @@ function concat<T>(a: T[] = [], b: T[] = []): T[] {
 }
 
 export const prop: string = program.prop;
-export const components: string[] = program.args;
+export const components: string[] = program.components;
 export const showProgress: boolean = program.progress;
 export const babelPlugins: ParserPlugin[] = concat(
   config.babelPlugins,
-  program.addBabelPlugin
+  program.babelPlugins
 );
 export const directory: string = program.directory || config.directory;
 export const gitignore: boolean = program.gitignore;